+++ packages/components/pages/f-checkout/test/component/f-checkout-delivery.component.spec.js
@@ -47,37 +47,5 @@ describe('f-checkout "delivery" component tests', () => {
                 expect(checkout.getFieldValue(field).length).toEqual(50);
             });
         });
-
-        describe('and age verification is required', () => {
-            beforeEach(() => {
-                // Arrange
-                checkout = new Checkout();
-                const checkoutInfo = {
-                    serviceType: 'delivery',
-                    isAuthenticated: true,
-                    isASAP: true,
-                    locale: 'en-AU',
-                    restrictions: 'age-restriction'
-                };
-                checkout.withQuery('&knob-Service Type', checkoutInfo.serviceType)
-                            .withQuery('&knob-Is User Logged In', checkoutInfo.isAuthenticated)
-                            .withQuery('&knob-Is ASAP available', checkoutInfo.isASAP)
-                            .withQuery('&knob-Locale', checkoutInfo.locale)
-                            .withQuery('&knob-Restrictions', checkoutInfo.restrictions);
-
-                checkout.loadAgeVerification();
-            });
-
-            it('should display a field error if the age is younger than 18', () => {
-                const todaysDate = new Date();
-                const dob = {
-                    day: todaysDate.getDay(),
-                    month: todaysDate.toLocaleString('default', { month: 'long' }),
-                    year: todaysDate.getFullYear() - 17 // The user is 17 years old
-                };
-
-                checkout.populateAgeVerificationForm(dob);
-            });
-        });
     });
 });